Add functioning for loop
This for loop only counts rounds via 'i'. roundNumber is not used and
has been commented out.

loop runs six times, but the first alert in the beginning of program
also runs, so the program overall asks for an answer submission
six times. Aim is to fix that with next commit.