Adjust logic so that game resets whenver it is impossible to win
Whenever there are two ties AND two computerWins or two ties AND
two userWins, the game declares the winner to be the player with
two wins since is impossible for the other player to catch up.
If 5 rounds have passed and neither computer or user have an advantage,
the game is declared a tie.